requirements.txt
streamlit>=1.35.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
plotly>=5.18.0
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
httpx>=0.27.0
altair>=5.0.0
